Refactor controllers into a separate subpackage
