+++ demo/src/main/java/com/example/demo/web/UserController.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.web;
 
 import com.example.demo.domain.ZetchUser;
 import com.example.demo.repo.UserRepository;
